// src/api/generated/model/internalCenterHttpHandlerNetworkConnectivityTestResponse.ts
/**
 * Generated by orval v8.5.2 🍺
 * Do not edit manually.
 * zvas API
 * zvas 初始化阶段 API 文档。
 * OpenAPI spec version: 1.0
 */
import type { InternalCenterHttpHandlerNetworkConnectivityTestData } from './internalCenterHttpHandlerNetworkConnectivityTestData';

export interface InternalCenterHttpHandlerNetworkConnectivityTestResponse {
  code?: number;
  data?: InternalCenterHttpHandlerNetworkConnectivityTestData;
  message?: string;
  trace_id?: string;
}