// src/api/generated/model/internalCenterHttpHandlerNetworkInterfaceItem.ts
/**
 * Generated by orval v8.5.2 🍺
 * Do not edit manually.
 * zvas API
 * zvas 初始化阶段 API 文档。
 * OpenAPI spec version: 1.0
 */
import type { InternalCenterHttpHandlerNetworkInterfaceConfigData } from './internalCenterHttpHandlerNetworkInterfaceConfigData';
import type { InternalCenterHttpHandlerNetworkInterfaceStatusData } from './internalCenterHttpHandlerNetworkInterfaceStatusData';

export interface InternalCenterHttpHandlerNetworkInterfaceItem {
  config?: InternalCenterHttpHandlerNetworkInterfaceConfigData;
  is_protected?: boolean;
  name?: string;
  role?: string;
  status?: InternalCenterHttpHandlerNetworkInterfaceStatusData;
}