// src/api/generated/model/internalCenterHttpHandlerNetworkInterfaceResponse.ts
/**
 * Generated by orval v8.5.2 🍺
 * Do not edit manually.
 * zvas API
 * zvas 初始化阶段 API 文档。
 * OpenAPI spec version: 1.0
 */
import type { InternalCenterHttpHandlerNetworkInterfaceItem } from './internalCenterHttpHandlerNetworkInterfaceItem';

export interface InternalCenterHttpHandlerNetworkInterfaceResponse {
  code?: number;
  data?: InternalCenterHttpHandlerNetworkInterfaceItem;
  message?: string;
  trace_id?: string;
}