// src/api/generated/model/internalHandlerSystemHealthResponse.ts
/**
 * Generated by orval v8.5.2 🍺
 * Do not edit manually.
 * zvas API
 * zvas 初始化阶段 API 文档。
 * OpenAPI spec version: 1.0
 */
import type { InternalHandlerSystemHealthData } from './internalHandlerSystemHealthData';

export interface InternalHandlerSystemHealthResponse {
  code?: number;
  data?: InternalHandlerSystemHealthData;
  message?: string;
  trace_id?: string;
}