// src/api/generated/model/internalHandlerSystemSettingsResponse.ts
/**
 * Generated by orval v8.5.2 🍺
 * Do not edit manually.
 * zvas API
 * zvas 初始化阶段 API 文档。
 * OpenAPI spec version: 1.0
 */
import type { InternalHandlerSystemSettingsData } from './internalHandlerSystemSettingsData';

export interface InternalHandlerSystemSettingsResponse {
  code?: number;
  data?: InternalHandlerSystemSettingsData;
  message?: string;
  trace_id?: string;
}