// src/api/generated/model/internalHandlerSystemVersionResponse.ts
/**
 * Generated by orval v8.5.2 🍺
 * Do not edit manually.
 * zvas API
 * zvas 初始化阶段 API 文档。
 * OpenAPI spec version: 1.0
 */
import type { InternalHandlerSystemVersionData } from './internalHandlerSystemVersionData';

export interface InternalHandlerSystemVersionResponse {
  code?: number;
  data?: InternalHandlerSystemVersionData;
  message?: string;
  trace_id?: string;
}